reorganized files for worker and add queue checking logic
